refactor: Params 클래스 패키지 structure로 이동
