+++ src/main/java/codesquad/server/http/HttpRequest.java
@@ -1,5 +1,6 @@
 package codesquad.server.http;
 
+import codesquad.server.structure.Params;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
