+++ src/test/java/codesquad/server/HttpServerTest.java
@@ -1,0 +1,10 @@
+package codesquad.server;
+
+import org.junit.jupiter.api.Test;
+
+public class HttpServerTest {
+    @Test
+    void 요청이_끝나면_socket을_잘_닫는지_확인() {
+    }
+
+}
